+++ 0-Coins/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".features.available_books.views.activities.CoinsActivity"
+            android:name=".ui.features.available_books.views.activities.OrderBookActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
