--- 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/features/available_books/domain/models/data/order_book/response/Bids.kt
@@ -1,9 +1,0 @@
-package com.axiasoft.android.zerocoins.features.available_books.domain.models.data.order_book.response
-
-import com.google.gson.annotations.SerializedName
-
-data class Bids(
-    @SerializedName("book"  ) var book          : String? = null,
-    @SerializedName("price" ) var price         : String? = null,
-    @SerializedName("amount") var amount        : String? = null,
-)
