--- 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/features/available_books/domain/models/data/order_book/response/ListOrderBookResponse.kt
@@ -1,8 +1,0 @@
-package com.axiasoft.android.zerocoins.features.available_books.domain.models.data.order_book.response
-
-import com.google.gson.annotations.SerializedName
-
-data class ListOrderBookResponse (
-    @SerializedName("asks"  ) var asks          : ArrayList<Ask>? = null,
-    @SerializedName("bids"  ) var bids          : ArrayList<Bids>? = null
-)
