+++ 0-Coins/app/src/main/java/com/axiasoft/android/zerocoins/ui/features/available_books/domain/mappers/BookOrderPropietiesMapper.kt
@@ -1,0 +1,16 @@
+package com.axiasoft.android.zerocoins.ui.features.available_books.domain.mappers
+
+import com.axiasoft.android.zerocoins.R
+
+fun bookOrderName(bookOrderCode: String): String{
+    return ""
+}
+
+fun bookOrderImageResId(bookOrderCode: String): Int{
+    return R.drawable.ic_no_image
+}
+
+fun getBookOrderType(bookOrderCode: String): CoinNameAndImage? {
+    return CoinNameAndImage.values().firstOrNull {it.coinKey == bookOrderCode}
+}
+
