+++ 0-Coins/app/src/main/res/layout/activity_coins.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".features.available_books.views.activities.CoinsActivity"
+    tools:context=".ui.features.available_books.views.activities.OrderBookActivity"
     android:id="@+id/cl_cointainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
