+++ 0-Coins/app/src/main/res/layout/book_coin_item.xml
@@ -7,7 +7,7 @@
     <data>
         <variable
             name="book"
-            type="com.axiasoft.android.zerocoins.features.available_books.domain.models.data.book.response.Book" />
+            type="com.axiasoft.android.zerocoins.ui.features.available_books.domain.models.data.book.response.Book" />
     </data>
 
     <androidx.cardview.widget.CardView
